curso-python-e-langchain/requirements.txt
openai==1.13.3
python-dotenv
langchain==0.1.11
langchain-openai==0.0.8